requirements.txt
blinker==1.7.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
flask~=3.0.3
fonttools==4.47.0
itsdangerous==2.1.2
jinja2==3.1.2
kiwisolver==1.4.5
markupsafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
sqlalchemy==2.0.24
typing-extensions==4.9.0
werkzeug==3.0.1
pytz~=2024.1
qrcode~=7.4.2
PyQRCode~=1.2.1
uuid~=1.30
python-dotenv~=1.0.1